+++ src/search/landmarks/landmark_count_heuristic.cc
@@ -37,7 +37,7 @@ LandmarkCountHeuristic::LandmarkCountHeuristic(const options::Options &opts)
           (!task_properties::has_axioms(task_proxy) &&
            (!task_properties::has_conditional_effects(task_proxy) || conditional_effects_supported))),
       successor_generator(nullptr) {
-    utils::g_log << "Initializing landmarks count heuristic..." << endl;
+    utils::g_log << "Initializing landmark count heuristic..." << endl;
 
     /*
       Actually, we should like to test if this is the root task or a
@@ -52,8 +52,17 @@ LandmarkCountHeuristic::LandmarkCountHeuristic(const options::Options &opts)
         utils::exit_with(utils::ExitCode::SEARCH_UNSUPPORTED);
     }
 
+    utils::Timer lm_graph_timer;
+    utils::g_log << "Generating landmark graph..." << endl;
     shared_ptr<LandmarkFactory> lm_graph_factory = opts.get<shared_ptr<LandmarkFactory>>("lm_factory");
     lgraph = lm_graph_factory->compute_lm_graph(task);
+    utils::g_log << "Landmark graph generation time: " << lm_graph_timer << endl;
+    utils::g_log << "Landmark graph contains " << lgraph->get_num_landmarks()
+                 << " landmarks, of which " << lgraph->get_num_disjunctive_landmarks()
+                 << " are disjunctive and " << lgraph->get_num_conjunctive_landmarks()
+                 << " are conjunctive." << endl;
+    utils::g_log << "Landmark graph contains " << lgraph->get_num_edges()
+                 << " orderings." << endl;
     bool reasonable_orders = lm_graph_factory->use_reasonable_orders();
     lm_status_manager = utils::make_unique_ptr<LandmarkStatusManager>(*lgraph);